Add low-level taproot helpers (#3086)
We add a new commitment format for simple taproot channels,
and implement creating and verifying musig2 and taproot sigs
for these channels.

Wire format and channel creation/update workflows have not
been modified, this will be done in another PR. This is currently
dead code, but is ready to be used and extensively tested!