Rework closing channel balance computation
We re-work the channel balance computation (`CheckBalance.scala`) used
to continuously monitor the overall balance of the node. We previously
assumed that all pending HTLCs would timeout on-chain, and monitored
our mempool to deduplicate unconfirmed transactions and RBF attempts.
But it is actually simpler to assume that all pending HTLCs will be
fulfilled, and keep counting them in our off-chain balance until the
spending transaction reaches `min-depth`.

We introduce a distinction between recently confirmed transactions and
deeply confirmed transactions in our on-chain balance: there will be
cases where we don't know yet whether we'll be able to gets funds back
and the information of pending on-chain funds lets us verify that our
overall balance is looking good. It always eventually converges to be
in our on-chain balance.

Fixes #3085