Stop storing commit tx and HTLC txs in channel data
We previously stored the commitment transaction and HTLC transactions
in our channel data along with the remote signatures, without our local
signatures. It is unnecessary to store those transactions, as we have
all the data we need to recompute them on-the-fly if we need to force
close.

Storing those transactions can easily use a few hundreds of kB per
channel, which adds up to the post-restart latency when channels need
to be read from the DB.

By not storing those transactions in our commitments, we also remove
the need to support encoding them, which gives us greater flexibility
to change their format in the future.

It also made it confusing in tests, where we sometimes use unsigned
transactions as if they were signed.