Distinguish unsigned HTLC txs from signed HTLC txs
There are two stages of HTLC transactions: we first generate unsigned
transactions to create our signatures for the remote commitment when
sending our `commit_sig`. Then we receive our peer's signatures for
the HTLC transactions for our local commitment, which we simply store.

We only need them when we want to force-close, at which point we enrich
HTLC transactions with their witness data (preimage for incoming HTLCs
and remote signatures).

We make this explicit in our transactions hierarchy, which lets us
simplify the publisher actors, get rid of the `redeemable` functions
in `ErrorHandlers` and clean up `Helpers.scala`.