Parse offers and pay offers with currency
The `payoffer` API requires an amount but the user has currently no way to read the amount from the offer.
We add `parseoffer` so that the user can read the offer amount (among other things).
Since the user needs to specify the amount they want to pay, we allow paying offers with a currency (other than BTC).