Separate anchor txs from local and remote commit
Now that we don't need to support encoding anchor txs in our channel
data, we can separate the anchor tx spending our local commitment from
the anchor tx spending a remote commitment.

This is a trivial refactoring that will allow us to go further and
enrich transaction with the corresponding commitment keys in the next
commit.