Add keys and commitment format to closing transactions
Now that we never store `ForceCloseTransaction` instances in our channel
data, we can enrich those types with runtime information such as the
keys necessary to sign those transactions.

We also include the commitment format, which is necessary to compute the
redeem info and otherwise had to be provided as argument everywhere.

The codecs changes are safe since we only use the input info and htlc_id
when reading old data.