Rework replaceable tx publisher
We refactor the replaceable tx publisher to remove the wrapper classes
for replaceable txs: we instead introduce a `WalletInputs` class and
directly provide signing helpers on the `ForceCloseTransaction`
instances.

We avoid modifying the internal transaction of a `ForceCloseTransaction`
instance, which was confusing: we instead return a `Transaction` from
the signing functions without modifying the "base" force-close tx.

This simplifies the `ReplaceableTxFunder` and paves the way for using
it for all transaction types that can be RBF-ed (main transaction, htlc
delayed transactions, penalty transactions, etc).