Don't update local params in `OpenChannelInterceptor`
We previously allowed plugins to update parts of our `LocalParams` in
the `OpenChannelInterceptor` flow: this was an attempt to allow adding
funds to a dual-funded channel that wasn't initiated by us and set
`max-htlc-value-in-flight` correctly (taking into account the added
funding amount).

This isn't used yet, and we will compute `max-htlc-value-in-flight`
later in the following commits, so we remove this option, which
simplifies the code.