Simplify force-close transaction signing and replaceable publishers (#3106)
* Update `ClaimHtlcSuccessTx` to always include preimage

We always have the preimage available when we create claim-htlc-success
transactions, so we can directly include it instead of the payment hash
now that we don't store those transactions on-disk anymore.

* Distinguish unsigned HTLC txs from signed HTLC txs

There are two stages of HTLC transactions: we first generate unsigned
transactions to create our signatures for the remote commitment when
sending our `commit_sig`. Then we receive our peer's signatures for
the HTLC transactions for our local commitment, which we simply store.

We only need them when we want to force-close, at which point we enrich
HTLC transactions with their witness data (preimage for incoming HTLCs
and remote signatures).

We make this explicit in our transactions hierarchy, which lets us
simplify the publisher actors, get rid of the `redeemable` functions
in `ErrorHandlers` and clean up `Helpers.scala`.

* Separate anchor txs from local and remote commit

Now that we don't need to support encoding anchor txs in our channel
data, we can separate the anchor tx spending our local commitment from
the anchor tx spending a remote commitment.

This is a trivial refactoring that will allow us to go further and
enrich transaction with the corresponding commitment keys in the next
commit.

* Add keys and commitment format to closing transactions

Now that we never store `ForceCloseTransaction` instances in our channel
data, we can enrich those types with runtime information such as the
keys necessary to sign those transactions.

We also include the commitment format, which is necessary to compute the
redeem info and otherwise had to be provided as argument everywhere.

The codecs changes are safe since we only use the input info and htlc_id
when reading old data.

* Rework replaceable tx publisher

We refactor the replaceable tx publisher to remove the wrapper classes
for replaceable txs: we instead introduce a `WalletInputs` class and
directly provide signing helpers on the `ForceCloseTransaction`
instances.

We avoid modifying the internal transaction of a `ForceCloseTransaction`
instance, which was confusing: we instead return a `Transaction` from
the signing functions without modifying the "base" force-close tx.

This simplifies the `ReplaceableTxFunder` and paves the way for using
it for all transaction types that can be RBF-ed (main transaction, htlc
delayed transactions, penalty transactions, etc).