Split channel params and commit params
We extract some of the channel parameters that are currently in our
`LocalParams` and `RemoteParams` classes into a `CommitParams` class.
While we currently simply create instances of that class based on our
`LocalParams` and `RemoteParams`, the goal is to later move those
values out of `LocalParams` and `RemoteParams` to clearly split params
that apply to the channel for its entire lifetime from params that can
be updated on a per-commitment basis.

This is mostly straightforward refactoring, but there is one detail
worth reviewing in-depth: we preivously used the `to_self_delay` from
the *remote* parameters when creating our *local* commitments. We now
instead initially set the `to_self_delay` in our *local* `CommitParams`
based on the remote `to_remote_delay`, which has been renamed to make
it more clear. This is a bit subtle, but hopefully will be better
contained once we split `CommitParams` from `ChannelParams` in a future
PR.

We also refactor our unit tests to simplify the creation of the "init"
messages. This should make it less painful whenever we have to update
these tests.