Fix flaky 0-conf watch-published event (#3124)
When using 0-conf, we wait for the transaction to appear in our mempool
before sending `channel_ready` or `splice_locked`, which ensures that
our inputs are correctly "locked" in our `bitcoind` wallet. We rely on
ZMQ to notify us about all transactions that appear in our mempool, but
it is unfortunately not providing strong guarantees that events will be
emitted, which means that in some cases we only "discover" the tx when
it is actually confirmed.

To make this more robust, we emit the `NewTransaction` event ourselves
when we successfully publish our funding transactions.