Refactor attribution helpers and commands (#3125)
* Refactor attribution decryption to prepare for trampoline

We update `Sphinx.scala` to shift hmacs differently to make it easier
to implement trampoline attribution in follow-up PRs.

* Refactor attribution data in `CMD_FAIL_HTLC` and `CMD_FULFILL_HTLC`

We group attribution data in a dedicated (optional) class for each
command instead of always adding more fields, which requires updating
a lot of test code whenever we need to change the contents.

We take this opportunity to add the trampoline received at field, and
update the pending commands DB.