Deduplicate inputs that have already been spent
Our on-chain utxos can be spent as soon as they have 1 confirmation.
When that happens, we still need to deduplicate the corresponding
utxo from our off-chain balance, but it doesn't appear in our list
of unspent utxos. We fix that by fetching the last blocks and scan
all utxos that have been spent in those blocks, which indicates that
they must be ignored from our off-chain balance. Once they reach min
depth, the off-chain balance won't include them because they will be
stored in the `irrevocablySpent` set in the channel data.