Add unconfirmed transaction pruning when computing closing balance (#3119)
Our node balance computation was simplified in #3096, but it is now too
simplistic during closing, where some of the balance may be duplicated
between the off-chain part and unconfirmed or recently confirmed
transactions.

We deduplicate those by not counting in our off-chain balance of closing
channels outputs that are spent by an unconfirmed or recently confirmed
transaction that we've included in our on-chain balance.

Fixes #3098