Re-work my_current_funding_locked reestablish
We introduce a `retransmit_flags` field to ask our peer to retransmit
`announcement_signatures` if we're expecting them: this way we don't
need to retransmit `splice_locked` on reconnection to trigger the
exchange of `announcement_signatures`. We don't need to retransmit it
to let our peer know that we've seen enough confirmations for the splice
either, since `my_current_funding_locked` implies that.

This allows us to completely remove retransmission of `splice_locked` on
reconnection, and also get rid of the `your_last_funding_locked` TLV,
which greatly simplifies the reconnection logic.

This will work with taproot channels since we'll simply provide nonces
in `channel_reestablish` when we need our peer to send announcement
signatures.

Note that we use a different TLV type to more easily allow migrating
existing users of the previous versions of the spec to the latest one.