Extract `CommitParams` to individual commitments (#3118)
We introduce a new channel codec version where:

- we extract per-commitment parameters (commitment format and params
  such as `dust_limit` which we may want to dynamically update in the
  future)
- we clean-up the splice iterations on the `Commitments` structure
- we remove the funding transaction from the confirmed local funding
  status (once confirmed, we don't need it anymore: it is wasting DB
  space for no good reason, we can get it from the blockchain)
- we add a `maxClosingFeerate_opt` to the closing state, to allow
  limiting the feerate used in RBF attempts for safe transactions
- we add the commitment format to interactive-tx parameters
- we make channel type feature non permanent

We add a mechanism to prevent the node from starting-up, unless an
override is provided. This provides an opportunity to iterate on codecs
while knowing that node operators won't be affected if we make changes
that are backwards-incompatible. This is meant to be temporary.