Allow non-initiator RBF for dual funding (#3021)
We previously only allowed the opener to RBF a dual-funded channel.
This is not consistent with splicing, where both peers can initiate
RBF. There is no technical reason to restrict the channel creation,
we can allow the non-initiator to RBF if they wish to do so.

The only subtlety is in the case where there is a liquidity purchase.
In that case we want the opener to be the only one allowed to RBF to
guarantee that we keep the liquidity purchase (since the initiator is
the only one that can purchase liquidity).