Re-encode channel data using v5 codecs (#3149)
We force a re-encoding of all channel data using v5 codecs when updating
eclair. This will allow us to remove support for previous codecs in a
later release. If users try to directly upgrade to a release that does
not contain the legacy codecs, we'll fail at start-up and tell them to
first update to the release that contains this migration code as an
intermediate step.

We plan to remove support for channels that aren't using anchor outputs,
so we add a warning in the logs and a command to help users close such
existing channels.