Add a signPsbt API method
This is intended to be used when eclair manages bitcoin core's wallet private keys.
In that case, the bitcoin wallet is watch-only and can create and fund transactions or psbts but not sign them.
This new API method will allow signing such transactions, which is more flexible than the existing `sendonchain` API call.